Movement-Updates
Playermovement has been fixed to allow for both double jump functionality and changing the model when moving left to right.
Player Prefab was edited with Angular Drag and Gravity to make it less Floaty when jumping and maneuvering in the air.
Tilemaps made in the future MUST UTILIZE the material within the materials folder labeled: AntiWallStick